5213: Barrel files refactor: meta/uuid
