+++ src/client/pages/Section/DataTable/Table/hooks/_parseTable.ts
@@ -4,7 +4,7 @@ import { Col as TypeCol } from 'meta/assessment/col'
 import { Cycle } from 'meta/assessment/cycle'
 import { Row as TypeRow, RowType } from 'meta/assessment/row'
 import { Table } from 'meta/assessment/table'
-import { UUIDs } from 'meta/uuid'
+import { UUIDs } from 'meta/uuid/uuids'
 
 import { ColHeader, ODPColHeader } from 'client/pages/Section/DataTable/Table/types'
