+++ src/client/pages/Section/Descriptions/NationalDataDescriptions/DataSources/hooks/useDataSourcesData.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 
 import { CommentableDescriptionName, DataSource } from 'meta/assessment/descriptionValue'
 import { SectionName } from 'meta/assessment/section'
-import { UUIDs } from 'meta/uuid'
+import { UUIDs } from 'meta/uuid/uuids'
 
 import { useCommentableDescriptionValue } from 'client/store/data/descriptions/hooks/descriptions'
 import { useIsDescriptionEditable } from 'client/store/user/hooks/auth'
