+++ src/client/store/login/actions/changePassword.ts
@@ -4,7 +4,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit'
 import axios from 'axios'
 
 import { ApiEndPoint } from 'meta/api/endpoint'
-import { UUIDs } from 'meta/uuid'
+import { UUIDs } from 'meta/uuid/uuids'
 
 import { NotificationActions } from 'client/store/ui/notification/actions'
 import { NotificationMessage } from 'client/store/ui/notification/state'
