+++ src/meta/cycleData/contacts/newContactNode.ts
@@ -3,7 +3,7 @@ import { NodeValue } from 'meta/assessment/node'
 import { ContactField } from 'meta/cycleData/contact/field'
 import { ContactNode } from 'meta/cycleData/contact/node'
 import { NodeExtType } from 'meta/nodeExt/nodeExt'
-import { UUIDs } from 'meta/uuid'
+import { UUIDs } from 'meta/uuid/uuids'
 
 type Props = {
   countryIso: CountryIso
