+++ src/meta/measurement/unit.ts
@@ -1,5 +1,5 @@
 import { UnitName } from 'meta/measurement/unitName'
-import { UUID } from 'meta/uuid'
+import { UUID } from 'meta/uuid/uuid'
 
 export type Unit = {
   name: UnitName
