+++ src/meta/uuid/uuids.ts
@@ -1,7 +1,5 @@
 import { v4 } from 'uuid'
 
-export type { UUID } from './uuid'
-
 export const UUIDs = {
   getUuid: v4,
 }
